Fix large encryption data
The cipher update was misused. It is intended to do a multi part encryption. It would make sense the use when working with InputStreams or BufferStreams, however, protobuf has all the bytes in memory already, so for the moment it is going to be simple to do a one stage encryption.

The encryption function could be replaced to something withing the cipher itself now, as it is basically duplicated.

However there is a bunch more of work in order to make the encryption safer, including Wrapped keys/assymetric encryption or even simple hmac to guarantee authenticity. Both will extend the function code.