upgrade to java 11
Simply enable java 11 for the project.
No code changes at this moment.

Recently while working with CompletableFutures, we suffered with limitations from java 8.

Considering that the code (or at least some part) tends to rely on java crypto architecture jca, it is extremely important to rely on a maintained up to date jdk.