chore: introduce spotbugs
Spotbugs seems to be one of the first to support jspecify which seems to be promising.

Also with really active maintenance and sounds like a replacement to findbugs.

The test source code is kept on the side for the time being as there is quite some auto generated code that has to be removed from the inspection as well as valid points to be worked on, however would like to focus on the Nullability.

There will be following commits/pr introducing jspecify and defining default nullability .